Trainning_AI/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
pydantic
python-multipart
requests
openai
